+++ afkak/client.py
@@ -24,10 +24,11 @@ from twisted.python.compat import nativeString
 from twisted.python.compat import unicode as _unicode
 
 from .common import (
+    BrokerResponseError,
     TopicAndPartition, FailedPayloadsError, BrokerMetadata,
     PartitionUnavailableError, LeaderUnavailableError, KafkaUnavailableError,
-    UnknownTopicOrPartitionError, NotLeaderForPartitionError, check_error,
-    DefaultKafkaPort, RequestTimedOutError, KafkaError, kafka_errors,
+    UnknownTopicOrPartitionError, NotLeaderForPartitionError, _check_error,
+    DefaultKafkaPort, RequestTimedOutError, KafkaError,
     NotCoordinatorForConsumerError, OffsetsLoadInProgressError, UnknownError,
     ConsumerCoordinatorNotAvailableError, CancelledError,
 )
@@ -243,7 +244,8 @@ class KafkaClient(object):
             return False
         return all(
             self.partition_fully_replicated(TopicAndPartition(topic, p))
-                 for p in self.topic_partitions[topic])
+            for p in self.topic_partitions[topic]
+        )
 
     def close(self):
         # If we're already waiting on an/some outstanding disconnects
@@ -384,7 +386,7 @@ class KafkaClient(object):
             log.debug("%r: c_m_resp: %r", self, c_m_resp)
             if c_m_resp.error:
                 # Raise the appropriate error
-                resp_err = kafka_errors.get(
+                resp_err = BrokerResponseError.errnos.get(
                     c_m_resp.error, UnknownError)(c_m_resp)
                 raise resp_err
 
@@ -561,7 +563,7 @@ class KafkaClient(object):
         out = []
         for resp in responses:
             try:
-                check_error(resp)
+                _check_error(resp)
             except (UnknownTopicOrPartitionError, NotLeaderForPartitionError):
                 log.error('Error found in response: %s', resp)
                 self.reset_topic_metadata(resp.topic)