+++ afkak/test/test_consumer_integration.py
@@ -2,26 +2,19 @@
 # Copyright 2015 Cyan, Inc.
 # Copyright 2018 Ciena Corporation
 
-import os
-
 import logging
 
-from nose.twistedtools import threaded_reactor, deferred
+from nose.twistedtools import deferred, threaded_reactor
 from twisted.internet.defer import inlineCallbacks, returnValue
-
-from afkak import (Consumer, create_message, )
-from afkak.common import (
-    ProduceRequest, ConsumerFetchSizeTooSmall,
-    OFFSET_EARLIEST, OFFSET_COMMITTED,
-    )
-from afkak.consumer import FETCH_BUFFER_SIZE_BYTES
-from .fixtures import ZookeeperFixture, KafkaFixture
-from .testutil import (
-    kafka_versions, KafkaIntegrationTestCase, async_delay,
-    random_string,
-    )
-
 from twisted.trial import unittest
+
+from .. import Consumer, create_message
+from ..common import (OFFSET_COMMITTED, OFFSET_EARLIEST,
+                      ConsumerFetchSizeTooSmall, ProduceRequest)
+from ..consumer import FETCH_BUFFER_SIZE_BYTES
+from .fixtures import KafkaHarness
+from .testutil import (KafkaIntegrationTestCase, async_delay, kafka_versions,
+                       random_string)
 
 log = logging.getLogger(__name__)
 
@@ -33,21 +26,10 @@ class TestConsumerIntegration(KafkaIntegrationTestCase, unittest.TestCase):
 
     @classmethod
     def setUpClass(cls):
-        if not os.environ.get('KAFKA_VERSION'):  # pragma: no cover
-            log.warning("WARNING: KAFKA_VERSION not found in environment")
-            return
-
-        # Single zookeeper, 3 kafka brokers
-        zk_chroot = random_string(10)
-        replicas = 3
-        partitions = 2
-
-        cls.zk = ZookeeperFixture.instance()
-        kk_args = [cls.zk.host, cls.zk.port, zk_chroot, replicas, partitions]
-        cls.kafka_brokers = [
-            KafkaFixture.instance(i, *kk_args) for i in range(replicas)]
-        # server is used by our superclass when creating the client...
-        cls.server = cls.kafka_brokers[0]
+        cls.harness = KafkaHarness.start(
+            replicas=3,
+            partitions=2,
+        )
 
         # Startup the twisted reactor in a thread. We need this before the
         # the KafkaClient can work, since KafkaBrokerClient relies on the
@@ -56,13 +38,8 @@ class TestConsumerIntegration(KafkaIntegrationTestCase, unittest.TestCase):
 
     @classmethod
     def tearDownClass(cls):
-        if not os.environ.get('KAFKA_VERSION'):  # pragma: no cover
-            log.warning("WARNING: KAFKA_VERSION not found in environment")
-            return
-
-        for broker in cls.kafka_brokers:
-            broker.close()
-        cls.zk.close()
+        cls.assertNoDelayedCalls()
+        cls.harness.halt()
 
     @inlineCallbacks
     def send_messages(self, partition, messages):
@@ -207,7 +184,7 @@ class TestConsumerIntegration(KafkaIntegrationTestCase, unittest.TestCase):
         big_consumer.stop()
         self.successResultOf(d)
 
-    @kafka_versions("0.8.1", "0.8.1.1", "0.8.2.1", "0.8.2.2", "0.9.0.1")
+    @kafka_versions("all")
     @deferred(timeout=15)
     @inlineCallbacks
     def test_consumer_restart(self):
@@ -260,7 +237,7 @@ class TestConsumerIntegration(KafkaIntegrationTestCase, unittest.TestCase):
         consumer.stop()
         self.successResultOf(start_d2)
 
-    @kafka_versions("0.8.2.1", "0.8.2.2", "0.9.0.1")
+    @kafka_versions("all")
     @deferred(timeout=15)
     @inlineCallbacks
     def test_consumer_commit_offsets(self):